Move Tx primitives to primitive library
