+++ lib/wallet/integration/src/Test/Integration/Scenario/API/Byron/Transactions.hs
@@ -36,7 +36,7 @@ import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..) )
 import Cardano.Wallet.Primitive.Types.TokenPolicy
     ( mkTokenFingerprint )
-import Cardano.Wallet.Primitive.Types.Tx
+import Cardano.Wallet.Primitive.Types.Tx.TxMeta
     ( Direction (..), TxStatus (..) )
 import Cardano.Wallet.Unsafe
     ( unsafeFromText )
