+++ lib/wallet/src/Cardano/Wallet/DB/Sqlite/Schema.hs
@@ -70,6 +70,7 @@ import qualified Cardano.Wallet.Primitive.Types.Coin as W
 import qualified Cardano.Wallet.Primitive.Types.TokenPolicy as W
 import qualified Cardano.Wallet.Primitive.Types.TokenQuantity as W
 import qualified Cardano.Wallet.Primitive.Types.Tx as W
+import qualified Cardano.Wallet.Primitive.Types.Tx.TxMeta as W
 import qualified Data.ByteString.Char8 as B8
 
 share
